Network: corrected confusion in unit conversion (bit/bytes, seconds/milliseconds)

+++ modules/cloudsim/src/main/java/org/cloudbus/cloudsim/network/datacenter/HostPacket.java
@@ -45,7 +45,7 @@ public class HostPacket {
 	/**
 	 * The length of the data being sent (in bytes).
 	*/
-	double data;
+	long data;
 
 	/**
 	 * The time the packet was sent.
